refactor(template-graphql): clean up template
