+++ packages/template-graphql/src/app.js
@@ -1,10 +1,7 @@
+import { Miss, render } from 'hops-react';
 import React from 'react';
-import { Route, Switch, Link } from 'react-router-dom';
-
-import { render, Miss } from 'hops-react';
-import { graphqlExtension } from 'hops-graphql';
-
-import { Home } from './home';
+import { Route, Switch } from 'react-router-dom';
+import Home from './home';
 
 const App = () => (
   <Switch>
@@ -13,6 +10,4 @@ const App = () => (
   </Switch>
 );
 
-export default render(<App />, {
-  extensions: [graphqlExtension()],
-});
+export default render(<App />);
